refactor: Update User model and related fields
Update the User model and related fields in the Prisma schema. Change the id field to use a string type with a default value generated by cuid(). Add the emailVerified field as a DateTime type. Rename the profile_img_url field to image. Add createdAt and updatedAt fields with default values. Also update the userId field in the UserDrug model to use a string type.