refactor: Update database schema to use Supabase URL and remove unused fields
This commit updates the database schema in the `schema.prisma` file. It replaces the `DATABASE_URL` environment variable with `SUPABASE_URL` and removes the `DIRECT_URL` variable, which was only used for Supabase. Additionally, it removes the `emailVerified`, `counsellingPointsVoiceLink`, `otherResources`, `acute`, and `chronic` fields from the `User` and `Drug` models, as they are no longer needed. This refactor improves the codebase by simplifying the schema and aligning it with the latest database push models.